// skinet.client/src/app/api/model/cartItem.ts
/**
 * SKINET.Server
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CartItem {
  productId: number;
  propductName: string | null;
  price: number;
  quantity: number;
  pictureUrl: string | null;
  brand: string | null;
  type: string | null;
}